requirements.txt
numpy==1.26.2
Pillow==9.2.0
Pillow==10.1.0
pygame==2.5.2
PyQt5==5.15.10
PyQt5_sip==12.13.0
tensorflow==2.10.0
tensorflow_gpu==2.10.0
tensorflow_intel==2.10.0
